add cluster-deux Separate out local startup code from connection manager, small adjustments. Comment by Andrey.
