+++ src/ReactiveDomain.Persistence/EventStore/ConnectionSettingsBuilderExtensions.cs
@@ -3,9 +3,9 @@ using EventStore.ClientAPI;
 
 namespace ReactiveDomain.EventStore
 {
-    public static class ConnectionSettingsBuilderExtensions {
+    internal static class ConnectionSettingsBuilderExtensions {
 
-        public static T IfVerboseLogging<T>(this T t, Func<bool> cond, Func<T, T> builder) where T : ConnectionSettingsBuilder {
+        internal static T If<T>(this T t, Func<bool> cond, Func<T, T> builder) where T : ConnectionSettingsBuilder {
             return cond() ? builder(t) : t;
         }
     }
